add ?next=url query redirect to student sign ups
