+++ curiositymachine/tests/tests.py
@@ -7,7 +7,6 @@ from django.core.urlresolvers import reverse
 from profiles.models import Profile, UserRole
 from ..middleware import UnderageStudentSandboxMiddleware, UnapprovedMentorSandboxMiddleware, LoginRequiredMiddleware, LoginRequired
 from ..views import root
-from ..views.generic import UserJoinView
 from challenges.models import Progress, Challenge, Example
 from ..helpers import random_string
 from curiositymachine import decorators
@@ -415,104 +414,6 @@ def test_challenge_access_decorator_redirects_other(rf):
     response = wrapped(request, challenge_id=1, username='named')
     assert not view.called
     assert response.status_code == 302
-
-def test_user_join_view_redirects_logged_in_user(rf):
-    request = rf.get('/whatever')
-    user = User()
-    profile = Profile(user=user)
-    request.user = user
-    view = UserJoinView.as_view(logged_in_redirect='/redirect')
-    assert type(view(request)) == HttpResponseRedirect
-    assert view(request).url == '/redirect'
-
-def test_user_join_view_sets_source_in_initial_form_data(rf):
-    request = rf.get('/whatever')
-    request.user = AnonymousUser()
-
-    m = mock.MagicMock()
-    view = UserJoinView.as_view(form_class=m)
-    view(request, source='source')
-
-    args = m.call_args
-    assert 'initial' in args[1]
-    assert 'source' in args[1]['initial']
-    assert args[1]['initial']['source'] == 'source'
-
-def test_user_join_view_sets_source_in_form_data_from_url(rf):
-    request = rf.post('/whatever', {'source': None})
-    request.user = AnonymousUser()
-
-    class SubUserJoinView(UserJoinView):
-        def form_valid(self, form):
-            pass
-
-    m = mock.MagicMock()
-    view = SubUserJoinView.as_view(form_class=m)
-    view(request, source='source')
-
-    args = m.call_args
-    assert 'data' in args[1]
-    assert 'source' in args[1]['data']
-    assert args[1]['data']['source'] == 'source'
-
-def test_user_join_view_strips_source_if_not_in_url(rf):
-    request = rf.post('/whatever', {'source': 'source'})
-    request.user = AnonymousUser()
-
-    class SubUserJoinView(UserJoinView):
-        def form_valid(self, form):
-            pass
-
-    m = mock.MagicMock()
-    view = SubUserJoinView.as_view(form_class=m)
-    view(request)
-
-    args = m.call_args
-    assert 'data' in args[1]
-    assert 'source' not in args[1]['data']
-
-def test_user_join_view_templates(rf):
-    request = rf.get('/whatever')
-    request.user = AnonymousUser()
-
-    m = mock.MagicMock()
-    view = UserJoinView.as_view(form_class=m, prefix='usertype')
-
-    response = view(request)
-    assert response.template_name == ['profiles/usertype/join.html']
-
-    response = view(request, source='somesource')
-    assert response.template_name == ['profiles/sources/somesource/usertype/join.html', 'profiles/usertype/join.html']
-
-def test_user_join_view_redirects_to_success_url(rf):
-    request = rf.post('/whatever', {})
-    request.user = AnonymousUser()
-
-    class SubUserJoinView(UserJoinView):
-        def create_user(self, form):
-            return mock.MagicMock()
-
-    m = mock.MagicMock()
-    view = SubUserJoinView.as_view(form_class=m, prefix='usertype', success_url='/success')
-
-    response = view(request)
-    assert isinstance(response, HttpResponseRedirect)
-    assert response.url == '/success'
-
-def test_user_join_view_redirects_to_welcome_url_from_form(rf):
-    request = rf.post('/join/thisone/', {'welcome': 'true'})
-    request.user = AnonymousUser()
-
-    class SubUserJoinView(UserJoinView):
-        def create_user(self, form):
-            return mock.MagicMock()
-
-    m = mock.MagicMock()
-    view = SubUserJoinView.as_view(form_class=m, success_url='/success')
-
-    response = view(request, source='thisone')
-    assert isinstance(response, HttpResponseRedirect)
-    assert response.url == '/welcome/thisone'
 
 @pytest.mark.django_db
 def test_signal_student_started_first_project():
